Add integration tests for PoDoFo functionality
